1.2.7 HP
Fix for the building systems & placement system.

Issue ( Resolved )
For some reason whatever input was thrown at a bounds check without being localized before it was thrown into the bounds check making whatever input was made into a out of bounds error. This is no longer the case, since we convert it and apply it correctly if I've done it correctly / understood it correctly. Also changed the debug spawn height.